USB: typec: add SPDX identifiers to some files
Due to these typec files being moved into the drivers/usb/ directory in
this tree, they missed the larger "add SPDX tags to all files" work.  So
add the correct SPDX license tag, based on the license text in the file
itself.

The SPDX identifier is a legally binding shorthand, which can be used
instead of the full boiler plate text.

Bug: 112400445
Change-Id: If92a18c8f0d17dd355cfda6c64ae93e442c6e3e4
Cc: Guenter Roeck <linux@roeck-us.net>
Cc: Hans de Goede <hdegoede@redhat.com>
Cc: Yueyao Zhu <yueyao.zhu@gmail.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Kate Stewart <kstewart@linuxfoundation.org>
Cc: Philippe Ombredanne <pombredanne@nexb.com>
Acked-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Acked-by: Rui Miguel Silva <rmfrfs@gmail.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: pumahsu <pumahsu@google.com>
Signed-off-by: Angga <angga@linuxmail.org>