LRNG: enforce max generate ops without reseed of full entropy
Add LRNG_DRNG_MAX_WITHOUT_RESEED that defines the maximum number of
generate operations that are allowed without a reseed that has full
entropy.

In case the limit is reached, the LRNG becomes non-operational which
implies that the blocking interfaces again start to block until the LRNG
is fully seeded again.

Note, if we have multiple NUMA nodes, only one DRNG will have
insufficient seed. Yet, we mark the entire LRNG as non-operational. This
is done due to the fact that this behavior should have almost zero
chance of ocurring. In addition, addition, if the DRNG is the first DRNG
which is identical to the lrng_drng_init instance, we cannot simply
leave the DRNG in the unseeded state, because per default if the LRNG
sees an unseeded DRNG it uses the lrng_drng_init as it is seeded first
during boot time. Therefore, it is easiest to simply mark the entire
LRNG as non-operational to avoid adding complex code that has almost
zero chance of being invoked.

Change-Id: I571e7a1befe414e6b4aa7ee316fae380c70a1120
Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Angga <angga@linuxmail.org>