LRNG: remove code duplication
The service function lrng_unset_fully_seeded is added which marks the
provided DRNG as not fully seeded any more. If the DRNG is equal to the
initial DRNG, the entire LRNG is set to non-operational which mandates
the LRNG to be fully seeded again.

The service function also checks if sufficient entropy is available to
immediately trigger a reseed.

The switch code uses the new service function too which implies that the
force_reseed of the DRNG instance is not set to true any more. That is
appropriate as the DRNG is not going to be used anyway until it is
reseeded. The reseed operation is performed automatically when
sufficient entropy is available.

Change-Id: Ib1fd81933a0b9c4d455cce2fbde25a4a9b3e3281
Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Angga <angga@linuxmail.org>