fs: sdfat: __sdfat_submit_bio_write() should always submit a WRITE
sdfat version 2.4.5 introduced wbc_to_write_flags() as a compat
function for legacy kernels. Unfortunately, as the legacy
__sdfat_submit_bio_write() path was adapted to take advantage of
wbc_to_write_flags(), it used the return value direcly as the
rw paramater to submit_bio(), and the default return value is 0
(READ). This becomes a problem later when write paths assert that
bio_data_dir() is WRITE (i.e., write bit of bio->bi_rw is set).

Pass a bitwise-or of WRITE and the return value of wbc_to_write_flags()
to submit_bio() on this legacy path. This logically unifies the
approach with the newer path here and directly with the legacy path of
__sdfat_submit_bio_write2() in mpage.c.

Fixes: 395d051fd0d6be91f2214b749033632c1d2b2072 ("fs: sdfat: Update to version 2.4.5")
Change-Id: I70aea5aa315039d83e64dd80e45c1750c295d8e6
Signed-off-by: Kevin F. Haggerty <haggertk@lineageos.org>
Signed-off-by: Angga <angga@linuxmail.org>