npu_v2: Process IRQ bottom half in an IRQ thread instead of a worker
Queuing a worker from hard IRQ context is forbidden on RT due to sleepable
spin locks, so we must fix up the queue_work() usage in the IRQ handler.

As it turns out, there's really no need to use a worker here to begin with,
so simply move the worker itself into a threaded handler to fix the RT
issue and also process IRQs at an elevated priority, as they should be.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[dereference23: Adapt for npu_v2]
Signed-off-by: Alexander Winkowski <dereference23@outlook.com>